Renamed files and components for consistency & clarity
